requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
spotipy
python-dotenv
